requirements.txt
python-gitlab~=4.4.0
PyGithub~=2.3.0
pandas~=2.2.2
tldextract~=5.1.2
requests~=2.31.0